requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
discord-protos==0.0.2
discord.py-self==2.0.1
frozenlist==1.5.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
protobuf==6.30.2
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
yarl==1.19.0
